created a separete file to prevent redundencies of code
